Base

All types

ParticleCorrelations.gls_pair_radial_fun — Method

gls_pair_radial_fun(pair_corr_distance::Function, a12::T; polynomial_order::Int, mesh_size::Int)

Return a function gls_fun. For any radial distances r1 and r2 we have gls = gls_fun(r1,r2) where gls is an array of the Legendre coefficients for the pair correlation.

Using mathematics, we have that such that $g(r_1,r_2,\cos \theta_{12}) = \sum_{\ell_1 =0} \frac{2\ell_1 + 1}{4\pi} gls[\ells+1] P_{\ell_1}(\cos \theta_{12})$, where $g(r_1,r_2,\cos \theta_{12})$ is the radially symmetric pair-correlation, so it depends only on the radial distances $r_1$ and $r_2$, and the angle between two position vectors $\theta_{12}$.

The function gls_fun is calculated from the function pair_corr_distance, where pair_corr_distance(sqrt(r1^2 + r2^2 - 2r1 * r2 * cos(θ12))) gives the pair correlation.

ParticleCorrelations.number_density — Method

number_density(s::Specie)

Gives the number of particles per unit volume. Note this is given exactly by N / V where V is the volume of the region containing the origins of all particles. For consistency, volume_fraction is given by N * volume(s) / V.

ParticleCorrelations.pair_radial_fun — Method

pair_radial_fun(pair_corr::Function, a12::T; polynomial_order::Int, mesh_size::Int)

Return a function pair_radial such that pair_radial(r1,r2, cos(θ12)) gives the pair correlation particles at the radial distances r1 and r2, with and angle of θ12 between them.

The function pair_radial is calculated from the function pair_corr_distance, where pair_corr_distance(sqrt(r1^2 + r2^2 - 2r1 * r2 * cos(θ12))) gives the pair correlation.

ParticleCorrelations.volume_fraction — Method

Returns the volume fraction of the specie.

ParticleCorrelations.DiscretePairCorrelation — Type

DiscretePairCorrelation

Represents the pair correlation between two types of species, which could be the same.

ParticleCorrelations.DiscretePairCorrelation — Method

DiscretePairCorrelation(particle_centres::Vector, R::T, MonteCarloPairCorrelation{Dim}())

Calculates the isotropic pair correlation from one configuration of particles. To use many configurations of particles, call this function for each, then take the average of the pair-correlation.

ParticleCorrelations.DiscretePairCorrelation — Method

DiscretePairCorrelation(s::Specie, pairtype::PercusYevick, distances::AbstractVector)

Generates a DiscretePairCorrelation for the specie s by using the Percus-Yevick approximation. This distribution assumes particles are distributed accoriding to a random uniform distribution, and that particles can not overlap.

ParticleCorrelations.HardMedium — Type

HardMedium

Represents a solid type of material. Solid particles can not overlap. Currently the package has no soft medium, but this can easily be added.

ParticleCorrelations.MonteCarloPairCorrelation — Type

MonteCarloPairCorrelation{Dim} <: PairCorrelationType

Currently only used to create pair-correlations for particles that are uniformly randomly placed, except they can not overlap.

ParticleCorrelations.PairCorrelation — Type

PairCorrelation

A type used to store a pair-correlation. This represents the calculated pair-correlation.

ParticleCorrelations.PairCorrelationType — Type

PairCorrelationType

A type used to specify what type of pair correlation is to be used. This is like a tag, or a option, to specify which pair correlation is wanted.

ParticleCorrelations.Specie — Type

Specie

Represents a set of particles which are all the same. The type of particle is given by Specie.particle and the volume fraction this specie occupies is given by Specie.volume_fraction.

We can use Specie.numberofparticles to specify the number of particles, otherwise for an infinite Specie.numberofparticles = Inf.

The minimum distance between any two particles will equal outer_radius(Specie) * Specie.separation_ratio.